Move Database-related classes to Axolotl namespace
This will be our version of Eloquent, which can be used outside of
the Jaxion framework to as a standalone ORM.
