--- tests/Stubs/MetaBase.php
@@ -1,8 +1,0 @@
-<?php
-namespace Intraxia\Jaxion\Test\Stubs;
-
-use Intraxia\Jaxion\Model\Base;
-
-class MetaBase extends Base {
-
-}
