+++ tests/Stubs/MetaModel.php
@@ -1,0 +1,8 @@
+<?php
+namespace Intraxia\Jaxion\Test\Stubs;
+
+use Intraxia\Jaxion\Axolotl\Model;
+
+class MetaModel extends Model {
+
+}
